% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{binary_transform}
\alias{binary_transform}
\title{binary transform}
\usage{
binary_transform(x)
}
\arguments{
\item{x}{A variable with two possible values.}
}
\value{
A vector with values in {0,1}.
}
\description{
Converts any variable with two possible values to a {0,1} binary
variable.
}
\examples{
## Convert a variable that takes values 'A' and 'B' to 0 and 1
x <- sample( c('A','B'), size = 10, prob = c(0.5,0.5), replace = TRUE )
binary_transform( x )
}
