% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.ClassDefinition.R
\docType{class}
\name{Bootstrap-class}
\alias{Bootstrap-class}
\title{Bootstrap}
\value{
Object of class Bootstrap
}
\description{
Bootstrap is a container class for bootstrap samples.
}
\section{Slots}{

\describe{
\item{\code{inbag}}{In-bag bootstrap sample.}

\item{\code{oob}}{Out-of-bag bootstrap sample.}
}}

\seealso{
\linkS4class{BootstrapStatistic}, \link{bootstrap}
}
