% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.ClassDefinition.R
\docType{class}
\name{BootstrapStatistic-class}
\alias{BootstrapStatistic-class}
\title{BootstrapStatistic}
\value{
Object of class BootstrapStatistic
}
\description{
BootstrapStatistic is a container class for bootstrap samples augmented with a computed statistic.
}
\section{Slots}{

\describe{
\item{\code{statname}}{The name of a (possibly user-defined) statistic to compute on the bootstrap sample.}

\item{\code{arglist}}{A list of arguments passed to the function referenced by statname.}

\item{\code{variable}}{The name of the variable on which to compute statname.}

\item{\code{inbag_stat}}{The value of statname for the in-bag bootstrapped sample.}

\item{\code{oob_stat}}{The value of statname for the out-of-bag bootstrapped sample.}
}}

\seealso{
\linkS4class{Bootstrap}, \link{bootstrap}
}
