% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_party.R
\docType{class}
\name{CTree-class}
\alias{CTree-class}
\title{CTree}
\value{
An object of class CTree
}
\description{
CTree is a container class for trees created by \link[party]{ctree}.
}
\section{Slots}{

\describe{
\item{\code{tree}}{An object of class \link[party]{BinaryTree-class} produced by
\link[party]{ctree}.}

\item{\code{data}}{Training data.}

\item{\code{parameters}}{Control parameters}
}}

\seealso{
\link[party]{ctree}, \link[party]{BinaryTree-class}
}
