% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSDT.ClassDefinition.R
\name{cutpoints}
\alias{cutpoints}
\title{Get distribution of cutpoints for subgroups.}
\usage{
cutpoints(object, subgroup = NULL, subsub = NULL)
}
\arguments{
\item{object}{An object of class TSDT}

\item{subgroup}{A string decscription of a subgroup (optional)}

\item{subsub}{A string description of a sub-subgroup (optional)}
}
\value{
A vector containing the subgroup cutpoints.
}
\description{
Get distribution of cutpoints for subgroups.
}
\seealso{
\link{TSDT}
}
