% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{datetime.LocalPrint}
\alias{datetime.LocalPrint}
\title{将UTC时间使用本地时间进行显示，实际存储还是UTC时间}
\usage{
datetime.LocalPrint(x, timeZone = Sys.timezone())
}
\arguments{
\item{x}{标准PISIXct格式日期}

\item{timeZone}{时区名，详细见timeZoneNames}
}
\value{
显示当天的时间格式，存储保持不变
}
\description{
将UTC时间使用本地时间进行显示，实际存储还是UTC时间
}
\examples{
datetime.LocalPrint(datetime('2018-02-16 12:31:47'),'PRC');
}
