% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{datetimeInput}
\alias{datetimeInput}
\title{用于输入标准的日期时间}
\usage{
datetimeInput(x)
}
\arguments{
\item{x}{标准的日期变量，}
}
\value{
返回
}
\description{
用于输入标准的日期时间
}
\examples{
datetimeInput("2010-12-13 15:30:30");
}
