% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{ddmmyyyy}
\alias{ddmmyyyy}
\title{用于输入日月年格式的数据}
\usage{
ddmmyyyy(x, sep = "-")
}
\arguments{
\item{x}{输入数据}

\item{sep}{分隔符}
}
\value{
反回标准Date数据
}
\description{
用于输入日月年格式的数据
}
\examples{
ddmmyyyy(01-02-2018);
}
