% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSDT_scoring_functions.R
\name{diff_restricted_mean_survival_time}
\alias{diff_restricted_mean_survival_time}
\title{diff_restricted_mean_survival_time}
\usage{
diff_restricted_mean_survival_time(data, scoring_function_parameters = NULL)
}
\arguments{
\item{data}{data.frame containing response data}

\item{scoring_function_parameters}{named list of scoring function control parameters}
}
\value{
Difference in restricted mean survival time across treatment arms.
}
\description{
Computes the difference in restricted mean survival time across
treatment arms.
}
\details{
Computes the restricted mean survival time for the treatment and
control arms and returns the difference.
}
\seealso{
\link{TSDT}, \link[survival]{Surv}, \link[survRM2]{rmst2}
}
