% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{function_parameter_names}
\alias{function_parameter_names}
\title{function_parameter_names}
\usage{
function_parameter_names(FUN)
}
\arguments{
\item{FUN}{The name of a function}
}
\value{
A character vector of function parameter names
}
\description{
Returns a character vector of the specified function's parameters
}
\examples{
## Define a function
example_function <- function( parm1, arg2, x, bool = FALSE ){
  cat( "This is an example function.\\n" )
}

## Return the function parameter names
function_parameter_names( example_function )
}
