% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDateSeries.R
\name{getDateSeries}
\alias{getDateSeries}
\title{生成标准的日期型时间序列}
\usage{
getDateSeries(startDate = Sys.Date(), days = 100)
}
\arguments{
\item{startDate}{开始日期，默认为当天}

\item{days}{天数，可以为正数或负数，不能等于0}
}
\value{
返回时间序列
}
\description{
生成标准的日期型时间序列
}
\examples{
getDateSeries(as.Date('2018-01-01'),days-100);
}
