% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeSeries.R
\name{getTimeSeries}
\alias{getTimeSeries}
\title{获取指定日期时间开始的时间序号，具体到秒}
\usage{
getTimeSeries(startDatetime = Sys.time(), seconds = 86400)
}
\arguments{
\item{startDatetime}{开始日期，带时间部分，UTC}

\item{seconds}{秒数}
}
\value{
返回标准UTC或CST的时间类型
}
\description{
获取指定日期时间开始的时间序号，具体到秒
}
\examples{
 getTimeSeries();
}
