% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSDT_scoring_functions.R
\name{hazard_ratio}
\alias{hazard_ratio}
\title{hazard_ratio}
\usage{
hazard_ratio(data, scoring_function_parameters = NULL)
}
\arguments{
\item{data}{data.frame containing response data}

\item{scoring_function_parameters}{named list of scoring function control parameters}
}
\value{
Hazard ratio across treatment arms.
}
\description{
Computes the hazard ratio across treatment arms using a CoxPH
model.
}
\seealso{
\link{TSDT}, \link[survival]{Surv},  \link[survival]{coxph},
}
