% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOperating.R
\name{mid}
\alias{mid}
\title{从字符中间进行测试}
\usage{
mid(x, startIndex, num_char)
}
\arguments{
\item{x}{字符串}

\item{num_char}{位数}

\item{start}{开始位置}
}
\value{
返回字符串
}
\description{
从字符中间进行测试
}
\examples{
mid('sdfdsfdsf',2,4);
}
