% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_party.R
\docType{class}
\name{MOB-class}
\alias{MOB-class}
\title{MOB}
\value{
An object of class MOB
}
\description{
MOB is a container class for trees created by \link[party]{mob}.
}
\section{Slots}{

\describe{
\item{\code{tree}}{An object of class \link[party]{BinaryTree-class} produced by
\link[party]{mob}.}

\item{\code{data}}{Training data.}

\item{\code{parameters}}{Control parameters}
}}

\seealso{
\link[party]{mob}, \link[party]{BinaryTree-class}
}
