% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{na2empty}
\alias{na2empty}
\title{na2empty}
\usage{
na2empty(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with NA values replaced with empty string.
}
\description{
Replace all instances of NA in character variable with empty
string.
}
\examples{
## Create character variable with missing values
ex1 <- c( 'A', NA, 'B', NA, 'C', NA )
ex1

## Replace NAs with empty string
ex1 <- na2empty( ex1 )
ex1
}
\seealso{
\link{unfactor}
}
