% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{partition}
\alias{partition}
\title{partition}
\usage{
partition(x, n)
}
\arguments{
\item{x}{Vector to partition.}

\item{n}{Number of (roughly) equally-sized groups}
}
\value{
A list of partitions of the vector x.
}
\description{
Partitions a vector x into n groups of roughly equal size.
}
\examples{
x <- 1:10
partition( x, 3 )
}
