% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOperating.R
\name{replaceStr}
\alias{replaceStr}
\title{用于字符串的替代}
\usage{
replaceStr(x, pattern, newStr)
}
\arguments{
\item{x}{字符向量}

\item{pattern}{字符或正则表达式}

\item{newStr}{新的替代字符串}
}
\value{
返回值
}
\description{
用于字符串的替代
}
\examples{
replaceStr('abcd','c','f');
}
