% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOperating.R
\name{right}
\alias{right}
\title{从右边截取指定位数的字符串}
\usage{
right(x, num_char)
}
\arguments{
\item{x}{签字文本}

\item{num_char}{签字位数}
}
\value{
返回字符串
}
\description{
从右边截取指定位数的字符串
}
\examples{
right(letters,3);
}
