% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOperating.R
\name{splitStr}
\alias{splitStr}
\title{将字符串进行拆分开}
\usage{
splitStr(x, pattern)
}
\arguments{
\item{x}{字符串}

\item{pattern}{可以使用字符串或正则表达式[]}
}
\value{
返回一个列表
}
\description{
将字符串进行拆分开
}
\examples{
split.str('afsdsdf,bbbfsdfds,sdfds',',');
}
