% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsample.ClassDefinition.R
\docType{class}
\name{Subsample-class}
\alias{Subsample-class}
\title{Subsample}
\value{
Object of class
}
\description{
Subsmaple is a container class for subsamples.
}
\section{Slots}{

\describe{
\item{\code{training}}{Training data.}

\item{\code{validation}}{Validation data.}

\item{\code{test}}{Test data.}
}}

\seealso{
\link{subsample}
}
