% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOperating.R
\name{subsetStr}
\alias{subsetStr}
\title{自动过滤符合条件的字符串,也称为字符串的子集或截取}
\usage{
subsetStr(x, pattern)
}
\arguments{
\item{x}{字符向量}

\item{pattern}{可以使用文本或表达式}
}
\value{
返回符合条件的结果
}
\description{
自动过滤符合条件的字符串,也称为字符串的子集或截取
}
\examples{
subsetStr(letters,'a');
}
