% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSDT.ClassDefinition.R
\docType{methods}
\name{summary,TSDT-method}
\alias{summary,TSDT-method}
\title{Summary function for class TSDT.}
\usage{
\S4method{summary}{TSDT}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{TSDT}.}
}
\value{
A data.frame containing the superior subgroups identified by TSDT.
}
\description{
Summary function for class TSDT.
}
\seealso{
\link{TSDT}
}
