% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSDT.ClassDefinition.R
\docType{class}
\name{TSDT-class}
\alias{TSDT-class}
\title{TSDT}
\value{
Object of class TSDT
}
\description{
TSDT is a container class for TSDT samples and metadata.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of parameters used in construction of TSDT samples.}

\item{\code{samples}}{Vector of \linkS4class{TSDT_Sample} objects.}

\item{\code{superior_subgroups}}{data.frame containing summary statistics for superior
subgroups}

\item{\code{cutpoints}}{An object of class  \linkS4class{TSDT_CutpointDistribution}.}

\item{\code{distributions}}{A list of distributions of TSDT statistics.}
}}

\seealso{
\link{TSDT}, \linkS4class{TSDT_Sample},
\linkS4class{TSDT_CutpointDistribution}
}
