% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSDT_CutpointDistribution.ClassDefinition.R
\docType{class}
\name{TSDT_CutpointDistribution-class}
\alias{TSDT_CutpointDistribution-class}
\title{TSDT_CutpointDistribution}
\value{
Object of class TSDT_CutpointDistribution
}
\description{
Implementation of TSDT_CutpointDistribution class. This class
continuous split variable. If the subgroup contains more than one split
variable a distribution of numeric cutpoints is collected for each continuous
split variable in the subgroup definition.
}
\section{Slots}{

\describe{
\item{\code{Cutpoints}}{An object of class \link[hash]{hash-class}}
}}

\seealso{
\link{TSDT}, \link[hash]{hash}
}
