% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSDT_Sample.ClassDefinition.R
\docType{class}
\name{TSDT_Sample-class}
\alias{TSDT_Sample-class}
\title{TSDT_Sample}
\value{
Object of class TSDT_Sample
}
\description{
TSDT_Sample is a container class containing the in-bag and
out-of-bag data from a subsampled or bootstrapped dataset. This container
class also contains a data.frame containing the parsed tree that is fit on the
in-bag data.
}
\section{Slots}{

\describe{
\item{\code{inbag}}{A data.frame containing in-bag data}

\item{\code{oob}}{A data.frame containing out-of-bag data}

\item{\code{subgroups}}{A data.frame containing a parsed tree}
}}

\seealso{
\link{TSDT}
}
